DefaultLegacyOpener: Avoid triggering ImageJ 1.x bug
There is a bug in at least version 1.49b of legacy ImageJ in that
the ImagePlus tries to cast the ImageWindow associated with a
CompositeImage to the StackWindow subclass. The assumption that
hyperstacks or composite images are displayed in a StackWindow is
wrong, however, if the stack size is one.

Therefore, much to this developer's chagrin, we have to continue the
inconsistency of returning sometimes composite images, sometimes not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>